separated alert dialog code from show dialog
